requirements.txt
appdirs==1.4.4
beautifulsoup4==4.11.2
cachelib==0.9.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==39.0.1
cs50==9.2.2
finnhub-python==2.4.16
Flask==2.1.2
Flask-Session==0.4.0
frozendict==2.3.5
geographiclib==1.52
geopy==2.2.0
gnureadline==8.1.2
googlefinance==0.7
greenlet==1.1.2
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
instagram-private-api==1.6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
multitasking==0.0.11
nsepython==0.0.973
numpy==1.24.2
pandas==1.5.3
prettytable==0.7.2
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.1
requests-toolbelt==0.9.1
scipy==1.10.0
six==1.16.0
soupsieve==2.4
SQLAlchemy==1.4.39
sqlparse==0.4.2
termcolor==1.1.0
urllib3==1.26.10
webencodings==0.5.1
Werkzeug==2.1.2
yfinance==0.2.12
zipp==3.8.0
